This might address issue #1, or it might not; going to run this build on my phone for a bit to see if it does then either release the new version to market or go back to the drawing board.
It looks like the error happens because multiple instances of the broadcastreceiever get spun up, which means that the first one correctly saves the previous ringer state, but the next one sees it as "silent", and saves that as the old state. This means that once it gets restored, it just gets set back to silent anyways and bam - bug.