fixes issue #1
The previous change fixed part of the problem, and this fixes the rest.

The receiver (I think erroneously) gets called when it is registered. This means that when the user starts the service for the first time, reboots, or when the service is destroyed and restarted by the system, their volume will be set. Aside from being unintended behavior, it also causes their "old mode" to be overwritten, which means that if they are already plugged in (and the ringer is muted), their old mode will be saved as silent and the ringer will *not* be turned back on when they unplug. To prevent this, I'm using a member variable to record if it is the first try or not.