finally fixes issue #3.
The issue at hand was actually caused by a bug in the Android OS, introduced in 2.3, and extant in 3.0 as well (see https://groups.google.com/d/msg/android-developers/2H-zkME9FB0/ZymfmZTDaAkJ). There was an offered workaround for the issue, which apparently works. Hopefully this problem doesn't come up again.

Removed all the service foreground stuff, as it was just a workaround and shouldn't be neccessary.